Day 8 of advent. converted to BigInteger

+++ src/main/java/helpers/intCodeProgram/Instruction.java
@@ -1,5 +1,6 @@
 package helpers.intCodeProgram;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import static helpers.intCodeProgram.Instruction.ParameterMode.*;
@@ -26,55 +27,55 @@ public abstract class Instruction {
 
     private void createParameterModes() {
         parameterModes = new ParameterMode[4];
-        parameterModes[1] = getParameterModes(opcode / 100 % 10) ;
-        parameterModes[2] = getParameterModes(opcode / 1000 % 10) ;
-        parameterModes[3] = getParameterModes(opcode / 10000 % 10) ;
-        }
+        parameterModes[1] = getParameterModes(opcode / 100 % 10);
+        parameterModes[2] = getParameterModes(opcode / 1000 % 10);
+        parameterModes[3] = getParameterModes(opcode / 10000 % 10);
+    }
 
     private ParameterMode getParameterModes(int mode) {
-        if(mode == 0) {
-        return POSITION;
+        if (mode == 0) {
+            return POSITION;
         }
-        if(mode == 1) {
-        return IMMEDIATE;
+        if (mode == 1) {
+            return IMMEDIATE;
         }
         // mode == 2
         return RELATIVE;
-        }
+    }
 
     public abstract int getNumberOfParametersAndOpcode();
 
     /**
      * @return new pointer value
      */
-    public abstract int run(List<Integer> memoryState);
+    public abstract int run(List<BigInteger> memoryState);
 
-    public int getValueOfParam(int paramPosition, List<Integer> memoryState) {
+    public BigInteger getValueOfParam(int paramPosition, List<BigInteger> memoryState) {
         return getValueOfParam(paramPosition, memoryState, false);
     }
 
-    public Integer getValueOfParam(int paramPosition, List<Integer> integerList, boolean isSetter) {
+    public BigInteger getValueOfParam(int paramPosition, List<BigInteger> integerList, boolean isSetter) {
         ParameterMode paramMode = parameterModes[paramPosition];
         if (paramMode == RELATIVE) {
             return getValueOfParamRelativeMode(integerList, startingIndex, paramPosition);
         }
-        if(paramMode == POSITION && !isSetter) {
+        if (paramMode == POSITION && !isSetter) {
             return getValueOfParamPositionMode(integerList, startingIndex, paramPosition);
         }
         // paramMode == IMMEDIATE
         return getValueOfParamImmediateMode(integerList, startingIndex, paramPosition);
     }
 
-    private Integer getValueOfParamImmediateMode(List<Integer> integerList, int startingIndex, int i) {
+    private BigInteger getValueOfParamImmediateMode(List<BigInteger> integerList, int startingIndex, int i) {
         return integerList.get(startingIndex + i);
     }
 
-    private Integer getValueOfParamPositionMode (List<Integer> integerList, int startingIndex, int i) {
-        return integerList.get(integerList.get(startingIndex + i));
+    private BigInteger getValueOfParamPositionMode(List<BigInteger> integerList, int startingIndex, int i) {
+        return integerList.get(integerList.get(startingIndex + i).intValue());
     }
 
-    private Integer getValueOfParamRelativeMode (List<Integer> integerList, int startingIndex, int i) {
-        return integerList.get(integerList.get(startingIndex + i + relativeBase));
+    private BigInteger getValueOfParamRelativeMode(List<BigInteger> integerList, int startingIndex, int i) {
+        return integerList.get(integerList.get(startingIndex + i + relativeBase).intValue());
     }
 
 }
