+++ src/main/java/helpers/intCodeProgram/instructions/InputInstruction.java
@@ -2,10 +2,11 @@ package helpers.intCodeProgram.instructions;
 
 import helpers.intCodeProgram.Instruction;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class InputInstruction extends Instruction {
-    public Integer input;
+    public BigInteger input;
 
     public InputInstruction(int opcode, int startingIndex, int relativeBase) {
         super(opcode, startingIndex, relativeBase);
@@ -17,8 +18,8 @@ public class InputInstruction extends Instruction {
     }
 
     @Override
-    public int run(List<Integer> memoryState) {
-        memoryState.set(getValueOfParam(1, memoryState, true), input); // input hardcoded for now
+    public int run(List<BigInteger> memoryState) {
+        memoryState.set(getValueOfParam(1, memoryState, true).intValue(), input); // input hardcoded for now
         return startingIndex + getNumberOfParametersAndOpcode();
     }
 }
