+++ src/main/java/helpers/intCodeProgram/instructions/OutputInstruction.java
@@ -2,10 +2,11 @@ package helpers.intCodeProgram.instructions;
 
 import helpers.intCodeProgram.Instruction;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class OutputInstruction extends Instruction {
-    private Integer output;
+    private BigInteger output;
 
     public OutputInstruction(int opcode, int startingIndex, int relativeBase) {
         super(opcode, startingIndex, relativeBase);
@@ -17,13 +18,13 @@ public class OutputInstruction extends Instruction {
     }
 
     @Override
-    public int run(List<Integer> memoryState) {
+    public int run(List<BigInteger> memoryState) {
         output = getValueOfParam(1, memoryState);
 //        System.out.print(getValueOfParam(1, memoryState) + " ");
         return startingIndex + getNumberOfParametersAndOpcode();
     }
 
-    public int getOutput() {
+    public BigInteger getOutput() {
         return output;
     }
 }
